// src/app/views/dashboard/Setup/manage-module/models/manage-module.ts
export class ManageModule {
  id?: number;
  ciclo?: string;
  especialidad?: string;
  modulo?: string;
  modulo_id?: string;
  docente?: string;
  laboratorio?: string;
  programa?: string;
  turno?: string;
  hora_inicio?: string;
  hora_fin?: string;
  vacantes?: number;
  seccion?: string;
  estado?: boolean;
  fecha_inicio?: string;
  fecha_fin?: string;
  turno_hora?: string;
}