// src/app/views/dashboard/Setup/tuition/models/tuition.ts
export class Tuition {
  id?: number;
  estudiante_id?: number;
  ciclo_academico_id?: number;
  especialidad_id?: number;
  modulo_id?: number;
  especialidad_modulo_id?: number;
  pago?: string;
  importe_pago?: string;
  voucher_numero_operacion?: string;
  voucher_img?: string;
  modalidad?: string;
  voucher_aprobado?: boolean;
  observacion?: string;
}